Test split codes
